code splitting in new GACPage
