+++ client/src/components/gacPage/modals/InformationalModal.jsx
@@ -3,8 +3,9 @@ import Modal from "react-bootstrap/Modal";
 import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
 
-import style from "../../pages/css/GacPage.module.css";
-import { fenixPhoto } from "../functions/dataTreatment";
+import style from '../../../pages/css/GacPage.module.css';
+import { fenixPhoto } from "../../functions/dataTreatment";
+import LoadSpinner from "../../../hooks/loadSpinner";
 
 export const CreateMoreInfoModal = ({ show, handleClose, username }) => {
   const [error, setError] = useState(null);
